requirements.txt
aiogram>=3.22.0
aiohttp>=3.12.15
aiohttp-jinja2>=1.6.0
jinja2>=3.1.4
pyjwt>=2.8.0
flask>=3.1.2
psutil>=7.1.1
pytelegrambotapi>=4.29.1
python-dotenv>=1.1.1
requests>=2.32.5
black>=24.0.0